network_WiFi_SimpleConnect: don't scrape hostapd logs for deauth
Today, we check that the AP never logs a deauth attempt from the client
during the run. This includes anything sent during initial
authentication. It appears clients might decide to send deauth
(erroneously?) during association and otherwise might have trouble that
requires retries. It does not seem to make sense to penalize clients for
this.

After that deauth check was added, we also added a check that the client
doesn't disconnect after the connection is stable, by watching `iw
event`. This should cover a superset of the deauth behaviors we care
about (e.g., if the client gives up and decides to deauth after we've
gotten this far, we'll notice).

Remove the deauth check to avoid false test failures.

BUG=chromium:483076
TEST=run the test

Change-Id: Ief8febb4ab5030902c090c5bd32a281418a6dfae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1495917
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>