Regenerate the mock owner public key and metric_{on,off} policies
At some point, the acceptable format for the owner key changed, in terms
of header sizes.  This can be seen in
commit d29ae33459d4 ("Updated policy_testserver to support
verification key signatures"), in the chrome repo.  Unfortunately,
the code in this directory was not updated.  As a result, if a user
had logged into a test machine at any point before
running platform_KernelErrorPaths (i.e. the machine has an owner),
the test machine gets powerwashed and the test fails.

The original code for generating the public key and the policy blobs is
found attached to this document:

https://sites.google.com/a/google.com/chromeos/archive/testing-tools-for-device-policies

Along with it are the instructions to generate the keys.  We follow
these instructions and commit the updated code.

TEST=platform_KernelErrorPaths.BUG passes after this.  Tried running
platform_KernelErrorPaths.ALL, but it has other issues.

BUG=chromium:966593
Change-Id: Id05e869bcb375e4a85a4f25949d462bcd7765959
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/1653840
Commit-Queue: Salman Qazi <sqazi@google.com>
Tested-by: Salman Qazi <sqazi@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>