autotest: Add a new script for syncing disks.
Occasionally, devices will randomly perform a power wash, wiping the
stateful partition, and losing python.  This happens most often
after tests that perform an EC reset, such as in mode_switcher.

With the sync done via RPC, each step of the sync was creating new
writes to sync, so the sync could never fully complete.  This script
will perform the whole sync as a single remote command, with the
details logged on the server (autoserv) side.

I'm separating the creation of the script (here) from commits that
start using it.

BUG=b:150334906
BUG=b:140625492
TEST=Apply the two following commits that use this script, then run firmware_RecoveryButton and suite:faft_bios.

Change-Id: I29c1544b7a62b65b49615570ebda025f01769780
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2097602
Tested-by: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>