graphics_HwOverlays: skip video if NV12 not supported, part 2
crrev.com/c/2079098 landed a provision to check if modetest
showed support for NV12; somehow this doesn't seem to work on
e.g. coral or celes, and it's because the |is_video| argument
to run_once is not set.

This CL fixes that by removing that parameter and simply using
  "video" in html_file
which is even better.

It also adds retain_output=True to the modetest command run,
since I noticed it's not collected in autotest logs and is a
nice piece of information.

BUG=chromium:1006887,chromium:1059465

TEST=test_that $IP graphics_HwOverlays.video on celes

Change-Id: I4fed11237d6fdfbc32bb34f646115cdb7b3b7d53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2090507
Commit-Queue: Ilja H. Friedel <ihf@chromium.org>
Tested-by: Ilja H. Friedel <ihf@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>