faft: Allow using script-based blocking sync instead of RPC.
This commit allows switching from using RPC calls for each step,
which causes new writes, to running the sync script as a single step.

The default is not to use the new technique, allowing gradual testing.

It will fall back to the old technique if the command fails.

BUG=b:150334906
BUG=b:140625492
TEST=Run firmware_RecoveryButton and suite:faft_bios several times.

Change-Id: I0be976847246d7490621c777ec2dac461525d562
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2096924
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Brent Peterson <brentpeterson@chromium.org>