Add extension .journal~ in addition to .journal as journal log file
This CL ignores files with .journal~ extension in addition to .journal
file when the tests collects the artifact files.

Also this adds the extension to the unshurinkable extension.

".journal~" is the extension for corrupted journal log file,
accorting to the man:
http://manpages.ubuntu.com/manpages/xenial/en/man8/systemd-journald.service.8.html

BUG=chromium:930226
BUG=chromium:1066706
TEST=none

Change-Id: Ia4b4063fffa5c6217ca122efaf61679c12df7435
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2094432
Tested-by: Brian Norris <briannorris@chromium.org>
Auto-Submit: Yoshiki Iguchi <yoshiki@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>