autotest: allow pcap to be configured as router
In our WiFi test cells, there are two router devices. By default, we
configure one of them as an actual router, and the other as a pcap. In
some circumstances, it's useful to be able to configure both as routers
(this gives us the ability to attenuate them separately in grover cell
setups). This change allows tests to bring up both router devices as
routers through the pcap_as_router flag.

BUG=None
TEST=network_WiFi_RoamDiagnostics (WIP)
     network_WiFi_RoamNatural (WIP)
     various network_WiFi_SimpleConnect
     network_WiFi_RoamFT.*

Change-Id: I324050bc89e6be0ebac9bf0cf0c63fcbb46233ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2103166
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>