tast: Build Go executables when SSP is unavailable.
When running Tast tests locally with test_that, user expectation
is to pick up local changes to the test code. To achieve this,
pass -build=true to the tast command when SSP is unavailable.

In automated testing SSP must be always enabled so this change
should not affect automated testing.

Note that we used to raise an exception when remote test bundle
dir is not found, but the check is removed because remote test
bundle does not necessarily exist.

BUG=chromium:1066384
TEST=test_that DUT tast.example
TEST=bin/python_venv -m server.site_tests.tast.tast_unittest

Change-Id: I90f64792dc8ba41e1b99272c68be42a8560a59d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2132073
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>