autotest: Move terminate_old to library and reuse it
The multimedia_xmlrpc_server process sometimes fails to start with
'address already in use' errors.  The FAFT xmlrpc server had code to
deal with terminating old leftover processes, so this commit moves it
into a library, and makes multimedia_xmlrpc_server call it as well.

This includes a couple of modifications to the process matching:
* Match based on script name, not based on exact command line.
* Check process name before checking command line, to avoid checking
  irrelevant or inaccessible processes.

BUG=b:152270235, b:154331738
TEST=Run suite:audio_basic and firmware_FAFTRPC.all

Change-Id: Ie301ac1861a56f68b93fd7bf6ee08362ee30468a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2135064
Tested-by: Dana Goyette <dgoyette@chromium.org>
Commit-Queue: Dana Goyette <dgoyette@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>