faft: firmware_ECChargingState: Test for full charge
The added test verifies power_supply_info reports 'Fully charged' when
the battery is full. This test will most likely be run together with
the whole faft_ec test suite with AC plugged, which means the battery
should've charged to full or close to full when this test is run.

Thus, even though we set the timeout as 3 hrs for the worse case where
the initial battery level is low, the test is mostly likely to finish
in a much shorter time.

BUG=b:151181037
TEST=run the test on Dratini board

Change-Id: I9bb246bec62cf1f5d57a6bd07203faac7fbf799f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2138598
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Tested-by: Philip Chen <philipchen@chromium.org>
Commit-Queue: Philip Chen <philipchen@chromium.org>