firmware_Cr50CheckCap: check ec uart availability matches cap state
We've changed up how the EC uart capabilities are used. Add a basic
ec_board check to the test to verify the EC uart works when expected.

This happens to trigger b/153198965, because the test will now use the
EC console when the capabiltiy is disabled and then try to use it when
it's enabled without rebooting cr50. A test specific to that bug may be
added later.

We should look into doing basic verification for all capabilities.

BUG=b:153198965
TEST=run with 4.28 and 4.27. 4.27 fails now.

Change-Id: Ib90df9fc29ff68fed27ac8b503fc7c23eaf7fe45
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2138602
Reviewed-by: Namyoon Woo <namyoon@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>
Tested-by: Mary Ruthven <mruthven@chromium.org>