network_RoamFT: always assume AP 0->1 roaming
Trying to determine which AP we're "starting" from isn't really helpful.

Generally, most of the roaming-decision impetus for this test is based
on the fact that Chrome OS clients estimate better throughput for 5GHz
than 2.4GHz APs for similar RSSI, so we're most likely only going to get
a 1-way roam (2.4 to 5) unless something else is going wrong (e.g.,
we're failing to connect to 5GHz).

This also masks certain errors -- for instance, if we happen to be
roaming already in the instant where we "check which AP we are currently
connected" to, we end up making an arbitrary choice on what the target
BSSID is.

Drop all this, and just assume the roam is always from AP 0 to 1. This
works out OK, because we always ensure that we initially connected to 0,
and we'll pass the test even if it already has roamed before we get
around to manually requesting scans to initiate roaming. The only
downside is that the "wait_for_roam()" metrics (printed in debug logs)
will be a little inaccurate.

Fix up a few comments and formatting while at it.

BUG=none
TEST=run it repeatedly

Change-Id: I5d792e8e369670d1c57bfcd36cc2a38f9a07ef74
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2149393
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>