network_WiFi_RoamFT: weed out a few more false-PASS cases
I've seen cases in local runs where we have time to:

(a) try to roam to the 2nd AP (via FT);
(b) associate, but fail in DHCP (due to a real DUT bug); and
(c) re-associate (not a true "roam") successfully to the 1st AP.

Because the first failed roam still reached the "associated" state (b),
we moved on to check L3 connectivity; however, we're not really checking
*which* AP we reach a fully-configured state with. So if (c) completes
quickly enough, we can still PASS erroneously.

Weed these out to some extent by
(1) ensuring the BSSID we end up on is the expected one; and
(2) ensuring there are no disconnections along the way while we're
    figuring out both (1) and our connectivity state.
This doesn't really button down all the cases in which we could be
bouncing around before this 'with' block, but it helps in many.

BUG=none
TEST=run it

Change-Id: I9da341dc10bb1f0ff99dc4a51ae6cbce7517969e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2149515
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>