firmware_WriteProtectFunc: Skip EC if no chrome_ec
The modified test runs functions that rely on the presence of a
chrome_ec. This causes failures on that do not have a chrome_ec, such
as drallion; see attached bug.
The test should only run EC functions if the DUT has a chrome_ec.

BUG=b:153470588
TEST=firmware_WriteProtectFunc on a board with a chrome EC (ex. coral)
TEST=firmware_WriteProtectFunc on a board with no chrome EC (ex. sarien)
TEST=firmware_WriteProtectFunc.dev on a board with a chrome EC
TEST=firmware_WriteProtectFunc.dev on a board with no chrome EC

Change-Id: Ieec39cfbd6ab5f58ecd952a236caca6eee5853db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2149746
Commit-Queue: Greg Edelston <gredelston@google.com>
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Philip Chen <philipchen@chromium.org>