FAFTSetup: Warn if fw-testing-configs not found
In the attached bug, we saw that the lab environment does not have the
fw-testing-configs/ project, which should be found at
autotest/server/cros/faft/fw-testing-configs/. This repo can be found in
local repo checkouts or in CQ builders, but not in the lab environment.

The CL which caused the attached bug has since been reverted
(https://crrev.com/c/2147940) so we do not have any direct feedback as
to whether fw-testing-configs can be found in the lab environment.
We need to introduce some kind of feedback so that we know whether any
potential fixes are successful.

This CL causes firmware_FAFTSetup to report a warning if
fw-testing-configs/ cannot be found.

BUG=b:153768975
TEST=firmware_FAFTSetup
TEST=firmware_FAFTSetup but garble line 100 so it can't find the dir

Change-Id: Ib614fc945b628cf09ab852447d03ad6d7b1211af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2149760
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>