Move network interruption code into tests that trigger updates.
The 2 tests that interrupt AU via network disconnection do so by running
a client test that disconnects and reconnects the network
(autoupdate_DisconnectReconnectNetwork.py). This test is called when the
update has already started, and the update can progress significantly in
the time it takes to start/return from the client test. This is causing
test flakes in the lab, and complete test failure on moblab in some
partner environments.

This change moves the network disconnection test
code into the client update_engine_test.py, and triggers the network
disconnection from within the tests that start the update. This way, we
don't have to transition between client and server test so much when the
update is already running. Moblab is also given some additional
consideration and the upper bound on progress when we trigger the
interruption has been lowered to account for the faster download speeds
we are sometimes seeing.

BUG=chromium:1071611
TEST=test_that chromeos6-row15-rack10-host1.cros autoupdate_ForcedOOBEUpdate.interrupt.network.full -b octopus --no-retry --autotest_dir ~/trunk/src/third_party/autotest/files --args="job_repo_url=http://100.115.168.195:8082/static/octopus-release/R83-13020.8.0/autotest/packages"
autoupdate_Interruptions.network.full

Change-Id: I8459a64456edb0c69779f4684390c3aa1a944565
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2153915
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>