FAFTSetup: Log if fw-testing-configs found
Earlier this week we added a warning in firmware_FAFTSetup if the
fw-testing-configs directory cannot be found:
https://crrev.com/c/2149760

Recent runs of firmware_FAFTSetup do not show the warning statement,
which could indicate that the recent Autotest drone push fixed the
issue. Or, it could indicate that Autotest is running an
older-than-expected version of firmware_FAFTSetup.

For the sake of being cautious, add a logging statement as well if the
fw-testing-configs directory CAN be found. This should indicate that it
would be OK to cut back over to fw-testing-configs.

BUG=b:153768975
TEST=firmware_FAFTSetup

Change-Id: Ic0f916dd85b05bd41dd71271c3d53fcc99706302
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2154112
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>