faft_pd: firmware_PDVbusRequest: Extend the test for s0ix/s3 and g3
This CL adds the subtests for suspend/shutdown modes. When the DUT is
in S0, VBUS is always on. But when the DUT is in suspend or shutdown
mode and the servo_v4 is in SNK mode, VBUS is off. This CL handles
this case.

BUG=b:153390824
TEST=Pass the test with control.suspend and control.shudown on Dratini

Change-Id: I9c4112cc79ec28641e2a11d424063ec986deb00c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2155115
Tested-by: Philip Chen <philipchen@chromium.org>
Commit-Queue: Philip Chen <philipchen@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>