faft_pd: firmware_PDPowerSwap: Extend the test for s0ix/s3 and g3
This CL adds the following tests:

(1) When the DUT is in S0ix/S3/G3, the DUT PD port should report
dualmode is not supported.

(2) When the DUT power state is S0 and the PD port power role is SRC
, even if the power state transitions to S0ix/S3, the DUT PD port
should keep sourcing.

(3) If the DUT PD port reports dualmode is supported in S0:
    (a) When the DUT power state is S0ix/S3 and the PD port power role
    is SRC, it should accept the power swap request from the PD tester.
    (b) When the DUT power state is S0ix/S3 and the PD port power role
    is SNK, it should reject the power swap request from the PD tester.

BUG=b:153390824
TEST=Pass the test with control, control.suspend, control.shudown
on Dratini

Change-Id: I748be3955259f3ff6aebbbd9b63582a4072f12e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2161899
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: Jett Rink <jettrink@chromium.org>
Tested-by: Philip Chen <philipchen@chromium.org>
Commit-Queue: Philip Chen <philipchen@chromium.org>