Use list to project dict_keys.
Python3 d.keys is not a list.
It seems like it is impossible to test this in chroot as the unittest
jumps into python2.7-config exe file which is not the interpreter. I
tested it in 2.7 and the unittest passes, which means at least the code
is not worse than it used to be.
Also see https://chromium-review.googlesource.com/c/chromiumos/platform/moblab/+/2157406
for a mirror fix in moblab.

BUG=b:154548704
TEST=See above

Change-Id: I9c5516a58fae4b6ad82962ad4372c6f3d1493a8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2160635
Reviewed-by: Keith Haddow <haddowk@chromium.org>
Reviewed-by: Allen Li <ayatane@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Daniil Lunev <dlunev@chromium.org>