autoupdate_ForcedOOBEUpdate: Fix check for update event in logs.
The code to check that we have rebooted at OOBE also does a check that
we have found the post-reboot update check in the new update_enging.log
file. However the code here checks only once and moves on even if it is
not found.

BUG=chromium:1073812
TEST=None. Simple change.

Change-Id: I0b2d32be2ecdcf94f1b2c40ab57b78a49decf28e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2162469
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>