FAFT: Soft cutover to fw-testing-configs
Previously, we attempted to give FAFT a hard cutover to the new
fw-testing-configs repo (https://crrev.com/c/2111694).
However, due to issues with finding the new fw-testing-configs repo in
the lab container, the attached bug arose, and we had to revert that CL
(https://crrev.com/c/2147940).

Since then, an Autotest drone push has occurred on 4/13, and we have
introduced logging to check whether fw-testing-configs can be found
(https://crrev.com/c/2154112). Based on recent Stainless logs, it looks
like it can (go/stainless-logs-with-fw-testing-configs-found).
So, it seems same to cut back over to the new fw-testing-configs repo.

However, out of an abundance of caution, this is a soft cutover. If
fw-testing-configs is found, then we read configs from there. Otherwise,
we fall back on the old configs/ directory (and so don't break all FAFT
tests).

If this works, then we will delete the old config/ directory after we
have ensured that the two directories are in-sync.

BUG=b:153768975
TEST=config_unittest.py
TEST=Any FAFT test. Check the INFO log to see which dir it uses for FAFT
configs. Check the DEBUG log to make sure it doesn't show the "No
platform config file found at %s. Using default." message.
TEST=Any FAFT test, after garbling fw-testing-configs to ensure that the
test will use the fallback directory. Check logs the same way as above.
TEST=firmware_FAFTSetup

Change-Id: I9a25ec8d9cfeb84a9a7c6963119b42addb8988ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2163189
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Brent Peterson <brentpeterson@chromium.org>
Reviewed-by: Kevin Shelton <kmshelton@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>