bluetooth: Remove the sleep before running stop_discovery
Currently, it takes about 5 seconds to find peers in test_discover_device,
with another time.sleep, the time between start_discovery and
stop_discovery in about 10 seconds, which is close to cause a race
condition described in b:154691540. This change removes the time sleep
as a workaround.

BUG=b:154691540
TEST=run quick sanity stress on bluez-next
     run quick sanity on bluez

Change-Id: Ia935cff5ea99eccf2274b2c1fbe5865a3d1c74e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2166159
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Commit-Queue: Sonny Sasaka <sonnysasaka@chromium.org>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>