cr50_test: use password in fast_open
The password should not be set if testlab mode is disabled. This could
happen though if cleanup behaves poorly or there is a test error. This
issue should not affect all tests that follow the bad test. Add a
recovery mechanism to fast_open. Use the password to open CCD and use
CCD reset to clear the password. CCD tests use fast_open to setup CCD at
the start of the test, so clearing the password here should protect all
CCD tests from failing because of this issue.

BUG=b:154927294
TEST=Set the password, disable testlab mode, and lock CCD verify
firmware_Cr50Open can still pass.

Change-Id: I0dfd0f816df408a346a87a1405453dd4e8fde6f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2165506
Tested-by: Mary Ruthven <mruthven@chromium.org>
Commit-Queue: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Dana Goyette <dgoyette@chromium.org>