Change update type to critical in autoupdate_UpdateFromUI
autoupdate_FromUI is failing a lot in the lab if the DUT has not gone
through OOBE yet, which often seems to be the case. Passing the
critical_update=True flag to nebraska ensures the UI update check won't
be ignored.
Also added screenshots for future debugging, since this test interacts
with the UI.

BUG=chromium:1073782
TEST=autoupdate_FromUI.full right after running
autoupdate_NonBlockingOOBEUpdate.delta

Change-Id: If2d1cf87d3f968566cc823c632d8d3c5e244d9e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2165510
Reviewed-by: David Haddock <dhaddock@chromium.org>
Tested-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>