faft: Ensure to return from run_shutdown_cmd() unpingable
There is a delay between the time when the shutdown command is run
on DUT and the time when DUT really shuts down.

So when we return from run_shutdown_cmd(), the DUT may still be
online, which may cause some weird race conditions if the test
developers are not careful.

Since we've cleanly disconnected RPCproxy in run_shutdown_cmd(),
we should just ensure DUT is down when we return from this function,
making this function less error-prone.

BUG=b:154516645, b:151021634
TEST=run the FAFT cases where run_shutdown_cmd() is called

Change-Id: I57d0433155cc861563544071a899f1524befcfc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2168819
Tested-by: Philip Chen <philipchen@chromium.org>
Reviewed-by: Chen-Tsung Hsieh <chentsung@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Philip Chen <philipchen@chromium.org>