Make reboot detection during OOBE AU more robust.
We currently check for a reboot when the DUT fails to execute
update_engine_client --status. This works most of the time but it can
sometimes fail if the update_engine is just slow and the DUT is not rebooting.

When we miss the reboot, the wait_for_oobe_update function will just
timeout. So I've removed the reboot detection code as is because if it
doesn't work 100% of the time it is not useful.

I've added a new check here that checks the 2 most recent
update statuses (should be FINIALIZING/NEED_REBOOT and IDLE) and that the
number of logs found in /var/log/update_engine/ has increased.

BUG=chromium:1073855
TEST=autoudpate_ForcedOOBEUpdate and interrupts

Change-Id: I6c41ac0f99263ace3ab93d0ca650b3d300f2d6ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2172408
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>