autotest: align faft_test_debug behaviors to faft-test
faft_test_debug is the debugging version of the faft-test pool.  Until
recently, the faft test pool spanned between production labs and the
firmware prototyping/debug section of one of the high-touch labs.
faft_test_debug pool was applied to devices that had been in the
faft-test pool in the prototyping/debug lab, so it makes sense that they
should have the same config as the faft-test pool, as they exist to
debug problems in the faft-test pool.

BUG=b:155424745
TEST=None (consistent with previous commit to this file)

Change-Id: Iaa74c276aa5cc106bec9921a86e65a563d222e0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2171946
Tested-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Kevin Shelton <kmshelton@chromium.org>