Make checking for update resuming more robust.
On both reboot interrupt tests we occasionally fail with 'update did not
continue where it left off'. Sometimes this is correct but a lot of
times it is due to how update_engine_client --status reports the
initial progress for a resumed update.

The first time we call it, it may report a very low progress
and then subsequently jump to the correct progress. To get around this
I've added extra checks against the update_engine.log looking for the
specific entries saying that we are resuming.

BUG=chromium:1073867
TEST=autoupdate_interruptions, autoupdate_ForcedOOBE.interrupt.reboot

Change-Id: I32e4077cee46874bb21d1fd7656e2ffa465f5305
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2176063
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>