Deprecate chromiumos_test_platform stage 1.
This is leftover from the days of Brillo when we were going to update
different types of devices. It just makes the code harder to follow imo
so I'm getting rid of it.

I've removed the code that spits out *some* update_engine.log on
failure. I think it is easier to debug by checking the full update_engine.log
file along with the hostlog and nebraska logs.

We need to leave some of the code for now. But it will be removed
shortly.

BUG=chromium:1078706
TEST=autoupdate_EndToEndTest

Change-Id: I6f342df5e1d2b8f4cc5ecdc54ff8c008cd11c6e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2184534
Commit-Queue: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>