servo: break out servo_Verification from servo_LabstationVerification
servo_LabstationVerification should be about gathering the dut/servo
pairs on a labstation, and then kicking off verification on each one
of them. The next step towards that is explicitly breaking out the
test-suite to be run against the dut/servo pair from the runner
scheduled against the labstation.

To that end servo_Verification is scheduled like traditional tests i.e.
the target is the DUT, and the labstation is the servo host. It simply
gathers all tests under the `servo_lab` suite, and runs them in a loop
not dissimilar to how servo_LabstationVerification was implemented.

The main advantage here is that now test development is decoupled from
the suite and the custom setup initially (converting the labstation from
target to servo host).

BUG=None

// This shows the test running in the current form with the labstation
// as the target.
TEST=test_that --autotest_dir . $GUADO
servo_LabstationVerification.local --args="dut_ip=$DRU"

// This leverages servo_Verification to run a verification with the dut
// as the target, as the dut/servo pair are already known.
TEST=test_that --autotest_dir . $KEVIN servo_Verification.local
--args="servo_host=$V3"

Change-Id: I44d36fb7e0bf56d940ba0c88119f4f78d179f33e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2184821
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>