site_linux_router: don't ping manually-specified router
It's arguably (?) nice to run a ping test against an
automatically-derived host (e.g., "dut-host" becomes "dut-host-router"),
because there's a chance the test-running user doesn't know that the
"-router" address is automatically derived.

This value proposition is much less when the user manually specifies the
address, via '--args=router_addr=...'.

(Arguably, this ping test is wasteful regardless and can be removed
altogether -- it's not currently present in the Tast port, for one...)

The real reason for this change: in WFH situations, I'm leveraging SSH
tunnels to access my not-sitting-in-the-office DUTs and APs. So instead
of 'router_addr=AN.IP.ADD.RESS', I'm using 'router_addr=localhost:PORT'.
Such a host is not pingable, obviously.

Drop the ping test for manually-specified addresses.

Similarly, drop the ping test for the pcap device altogether. The
current implementation of the ping test for the pcap device is actively
unhelpful since it silently skips creating the pcap device which
results in unintuitive failures later. We should attempt to create the
device regardless of pingability, and rely on the more helpful
connection failure output from the later SSH attemtps.

BUG=none
TEST=`test_that [...] --args=router_addr=localhost:2223 \
      localhost:2222 suite:wifi_perf`

Change-Id: Ie41dd452933d1cb490d72dd8a7ab11fd28dc7bd7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2188814
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>