bluetooth: Recover from bluetoothd crashes
It's possible for bluetoothd to crash and affect subsequent tests as
a result. In the quick test start and end functions, add a check to make
sure the bluetoothd proxy object is still valid and wait for bluetoothd
to start if it is not.

BUG=b:155300051
TEST=Ran AdapterLESanity, killing bluetoothd during tests and watched
     as it recovered gracefully for the next test.

Change-Id: Ifb3555c60aa29da6944fad64ab572823e176809b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2188350
Commit-Queue: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Tested-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>