dev_manifest: Check out fw-testing-configs
The fw-testing-configs project is used by FAFT, and is needed for
firmware provisioning (provision_FirmwareUpdate).

Currently, the container which runs provisioning cannot find
fw-testing-configs, which is causing firmware provisioning to fail: see
the attached bug.
With this change, fw-testing-configs would be added to the container
after a new image push.

BUG=b:155797809
TEST=None

Change-Id: Ice584e4bf3ee4b7d3e3665991b368c9f5e04b054
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2188821
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Anh Le <anhdle@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>