bluetooth: Ignore differences caused by bluez-next
In bluez 5.54, it has its own battery profile so it is not able to be
discovered through dbus api, and that breaks the
le_gatt_client_attribute_browse_test.

In this change, we ignore the difference caused by above commit so DUT
running bluez or bluez-next can pass the test.

BUG=b:155364847
TEST=run bluetooth_AdapterLESanity.le_gatt_client_attribute_browse_test
with DUT running bluez/bluez-next

Change-Id: Iafe7052701a25a2a28689514b395db6b0fd8c18c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2189996
Tested-by: Yun-Hao Chung <howardchung@google.com>
Tested-by: Sonny Sasaka <sonnysasaka@chromium.org>
Reviewed-by: Sonny Sasaka <sonnysasaka@chromium.org>
Commit-Queue: Sonny Sasaka <sonnysasaka@chromium.org>