Remove code that assumes payloads mayd come from gs://chromeos-image-archive.
This test hasn't worked with images from this other gs bucket for a long
time and there are still a few remants in the code. The recent changes
to remove cros_au made assumptions that we are getting payloads from
gs://chromeos-releases/ only so just update the code here to be easier.

BUG=chromium:1078701
TEST=autoupdate_E2E

Change-Id: I06812272c1673f2de88319016afba31b7e1d744f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2192046
Tested-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>