bluetooth: Remove two role tests with known failure
There is a known block in kernel that prevents a new LE connection from
being established if the controller is already in a connection as slave
role. This condition is tested in two tests, putting red in the
dashboard. This CL disables the two tests until the fix lands.

BUG=b:153067387
TEST=None

Change-Id: I9ade303bfe019804e280d7a07a0d323677504135
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2191988
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>