Change devserver choose comments.
Previously we tried to find the least loaded devserver. Then creating
new processes to query all devservers was causing SIGTERM in the lab so
we changed it to just use resolve(). Since this is what every other test
does that finds a devserver and there is no plan for the infra team to
fix the SIGTERM issue lets just go ahead with the current approach.

BUG=chromium:1078701
TEST=au_e2e

Change-Id: I6edd2535699806c9aaaec62f6ca83aa10ab20159
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2192852
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>