Rewrite hostlog parsing in verify_update_events().
The AU tests get hostlogs from two different places: one from nebraska
and one from parsing the update_engine.log. The nebraska way returns a
hostlog with event_types (13,14,3,54) as integers. Parsing
update_engine.log returns them as strings.

The test code currently uses strings in the expected events. So when we
are using a nebraska hostlog, we read the json and then make a new python
dictionary for each event and convert each int value to string.
This seems kind of wasteful and makes the code more
complicated. We can change update_engine_event to use integers and skip
the hostlog conversion and just use the data returned by json.loads()

This means that the hostlog from parsing the update_engine.log will
need to be changed but it is easy since it is in the same class.

Secondly the code currently reads one event at a time from the hostlog.
This is left over from when we would compare the hostlog while the
update was in progress. Now that we are comparing at the end we can just
iterate through the actual and expected together.

BUG=chromium:1077490
TEST=au_e2e in mismatched/error/timeout mode. au_forcedoobe

Change-Id: Ib43929f409da49ab02dbf36f75cfdd4c577587cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2197294
Tested-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>