servo_LabstationVerification: parallel_simple implementation
This change is to enable running multiple servo_Verifications in
parallel on a labstation (and mimic a lab environment more closely).
It does so by storing a map of hostnames to dut_hosts. Those then are
accessed in the parallel_simple implementation in the devices job.
The result is that the tests run in parallel and each get their own
output directory.
This also removes the need to use subdir tagging as each host machine
gets their own logging.

BUG=b:156104080

// patchset 1 had a hard-coded second host to make this work. Please
// take a look there as to why this test is meaningful.
TEST=test_that --autotest_dir . $V3 servo_LabstationVerification.local
     --args="dut_ip=$KEVIN"
     ls /tmp/test_that_latest/results-1-*/
// observe a folder for each ip address passed in

Change-Id: I74370e6e873730310124a6fe173a6e723db8ec97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2209919
Reviewed-by: Garry Wang <xianuowang@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>