Save the logs in /mnt/stateful_partition/unencrypted/preserve/log.
These are the logs for the DUT state from before we wipe the stateful
partition with the source update.

Currently the control file for this test adds this dir as a "loggable"
to sysinfo. Sysinfo then collects the dir. However, it collects the
loggable each time a control file is executed on the DUT. So we end up
collecting the logs multiple times for both login tests, the main test,
and again for the iteration hooks. I verified that they are just copies
of the same files. It doesn't matter if you collect them right after
source update or right at the end of the test.

I looked into adding a field to the loggable "skip_client_tests" but it
would require changes to base_sysinfo, site_sysinfo, autotest,
server_job and it would be passing state around that doesn't really
belong in any of those classes. And I don't want to rewrite them.

Another issue is the list of "loggables"
(list of directories to save) is sent to the DUT each time a control
file is executed (sysinfo.before, sysinfo.before.iteration,
login_loginSuccess, sysinfo.after.iteration, sysinfo.after) and the
sysinfo state is sent to the DUT and the state is read back
(base_sysinfo.serialize() and deserialize()) so you would need to
remove the loggable "for some control files executed on the DUT" and
then add it back in afterwards.

So after digging into all of that it is way easier to just grab the
logs we want once in cleanup and store it in a directory that makes
sense for the test.

BUG=chromium:1078295
TEST=au_e2e locally

Change-Id: I19e7bf0f8456e380db54725c0a4f1ec7dcf28c83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/2210224
Tested-by: David Haddock <dhaddock@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>
Reviewed-by: Amin Hassani <ahassani@chromium.org>